+++ staticx/archive.py
@@ -79,6 +79,12 @@ class SxArchive(object):
 
         self.tar.addfile(t)
 
+    def add_fileobj(self, name, fileobj):
+        logging.info("Adding {}".format(name))
+        tarinfo = self.tar.gettarinfo(arcname=name, fileobj=fileobj)
+        tarinfo.mode = make_mode_executable(tarinfo.mode)
+        self.tar.addfile(tarinfo, fileobj)
+
     def add_program(self, path, name):
         """Add user program to the archive
 
@@ -102,7 +108,7 @@ class SxArchive(object):
         # Store a link to the program so the bootloader knows what to execute
         self.add_symlink(PROG_FILENAME, name)
 
-    def add_library(self, path):
+    def add_library(self, path, exist_ok=False):
         """Add a library to the archive
 
         The library will be added with its base name.
@@ -110,6 +116,8 @@ class SxArchive(object):
         """
 
         if basename(path) in self._added_libs:
+            if exist_ok:
+                return
             raise LibExistsError(basename(path))
 
         # 'recursively' step through any symbolic links, generating local links inside the archive