+++ staticx/hooks/pyinstaller.py
@@ -16,7 +16,7 @@ def process_pyinstaller_archive(ctx):
 
     # Attempt to open the program as PyInstaller archive
     try:
-        pyi_ar = CArchiveReader(ctx.program)
+        pyi_ar = CArchiveReader(ctx.orig_prog)
     except:
         # Silence all PyInstaller exceptions here
         return
@@ -105,13 +105,9 @@ class PyInstallHook(object):
         for deppath in deps:
             dep = os.path.basename(deppath)
 
-            if dep in self.sx_ar.libraries:
-                logging.debug("{} already in staticx archive".format(dep))
-                continue
-
             if self.pyi_ar.toc.find(dep) != -1:
                 logging.debug("{} already in pyinstaller archive".format(dep))
                 continue
 
             logging.debug("Adding {} to archive".format(dep))
-            self.sx_ar.add_library(deppath)
+            self.sx_ar.add_library(deppath, exist_ok=True)